% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_output.R
\name{addUrlLink}
\alias{addUrlLink}
\title{Create a html balise for Gene Ontology description on ebi.ac.uk}
\usage{
addUrlLink(name)
}
\arguments{
\item{name}{charactor of Gene Ontology name.}
}
\value{
charactor of html balise
}
\description{
Create a html balise for Gene Ontology description on ebi.ac.uk
}
