% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server part of IIDEA}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{automatic inputs of the shiny server}

\item{output}{automatic output}

\item{session}{automatic session}
}
\value{
a server instance
}
\description{
Server part of IIDEA
}
