% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_post-hoc.R
\name{boundGroup2}
\alias{boundGroup2}
\title{IIDEA post-hoc bound table on gene set.}
\usage{
boundGroup2(object)
}
\arguments{
\item{object}{An object of class 'SansSouci'}
}
\value{
a data.frame containing:
\itemize{
\item the name of the gene set
\item the number of genes in the gene set
\item the estimate TP bound
\item the estimate FDP bound
}
}
\description{
IIDEA post-hoc bound table on gene set.
}
