% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{cleanGo.GS}
\alias{cleanGo.GS}
\title{Clean go.gs data sets. Not used}
\usage{
cleanGo.GS(go.gs)
}
\arguments{
\item{go.gs}{}
}
\value{
a list of cleaned data set
}
\description{
Clean go.gs data sets. Not used
}
