% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{cleanMatrix}
\alias{cleanMatrix}
\title{Clean expression matrix gived by user}
\usage{
cleanMatrix(matrixFunc)
}
\arguments{
\item{matrixFunc}{a matrix for expression genes}
}
\value{
a list with the cleaned matrix, a text for error, a bool for error and the color of the error
}
\description{
Clean expression matrix gived by user
}
\details{
Clean the user matrix for IIDEA requirement. If the matrix does not
respect the requirement, an error is signaled. The color of the message means:
red: the matrix can be used
orange: the matrix doesnt respect but IIDEA update it and can be used
blue: everything is ok. Thanks ;)
}
