% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_output.R
\name{exampleData}
\alias{exampleData}
\title{Create example data set}
\usage{
exampleData()
}
\description{
Create example data set
}
\details{
Create example data set from the package sanssouci.data.
This dataframe aims at exporting in IIDEA (not to be used by IIDEA as public
data set)
}
