% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_bulkRNAseq_datasets}
\alias{load_bulkRNAseq_datasets}
\title{Install GSEAbenchmarkeR dataset from the TCGA RNA-seq compendium to be used in IIDEA}
\usage{
load_bulkRNAseq_datasets(
  names = c("BRCA", "HNSC", "KICH", "KIRC", "KIRP", "LUSC", "PRAD", "STAD", "UCEC")
)
}
\arguments{
\item{names}{vector of character of GSEAbenchmarkeR expression data
compendium.  Should be a subset of c("BRCA", "HNSC", "KICH", "KIRC", "KIRP",
"LUSC", "PRAD", "STAD", "UCEC")}
}
\description{
Install GSEAbenchmarkeR dataset from the TCGA RNA-seq compendium to be used in IIDEA
}
