% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_microarray_datasets}
\alias{load_microarray_datasets}
\title{Install GSEAbenchmarkeR dataset from the GEO2KEGG microarray compendium to be used in IIDEA}
\usage{
load_microarray_datasets(
  names = c("GSE14762", "GSE15471", "GSE18842", "GSE19728", "GSE5281_EC", "GSE23878",
    "GSE7305", "GSE3467", "GSE9476", "GSE38666_epithelia")
)
}
\arguments{
\item{names}{a vector of character of GSEAbenchmarkeR expression data
compendium.  Should be a subset of c("GSE14762", "GSE15471", "GSE18842",
"GSE19728", "GSE5281_EC","GSE23878", "GSE7305", "GSE3467","GSE9476",
"GSE38666_epithelia")}
}
\description{
Install GSEAbenchmarkeR dataset from the GEO2KEGG microarray compendium to be used in IIDEA
}
