% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{matchMatrixBiofun}
\alias{matchMatrixBiofun}
\title{Check if genes in gene set match with a gene set list}
\usage{
matchMatrixBiofun(geneNames, biofun)
}
\arguments{
\item{geneNames}{vector of genes names (vector of charactors)}

\item{biofun}{matrix for gene set}
}
\value{
list for warnings
}
\description{
Check if genes in gene set match with a gene set list
}
\details{
Check if there is at least one gene in gene set match with genes
in expression matrix.  The color of the message means:
green: non genes match
blue: everything is ok.
}
