% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{readCSV_sep}
\alias{readCSV_sep}
\title{Read multiple csv configuration}
\usage{
readCSV_sep(file, ...)
}
\arguments{
\item{file}{a character of the csv file path}

\item{...}{other parameter for read.csv function}
}
\value{
a dataframe containing the dataset from the csv file
}
\description{
Read multiple csv configuration
}
