% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{RNAseq_blca_GO}
\alias{RNAseq_blca_GO}
\title{Build gene set for BLCA RNAseq data set from sanssouci.data}
\usage{
RNAseq_blca_GO(RNAseq_blca)
}
\arguments{
\item{RNAseq_blca}{matrix}
}
\value{
matrix
}
\description{
Build gene set for BLCA RNAseq data set from sanssouci.data
}
