% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_post-hoc.R
\name{t_linear}
\alias{t_linear}
\title{Reference families}
\usage{
t_linear(alpha, k, m)
}
\arguments{
\item{alpha}{numeric value in (0,1)}

\item{k}{integer value in \verb{[1,m]}}

\item{m}{integer value}
}
\description{
Imported from sanssouci package
}
\examples{
m <- 10
alpha <- 0.05
thr <- t_linear(alpha, 1:m, m)
}
