% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_output.R
\name{thrYaxis}
\alias{thrYaxis}
\title{Create a new y-axis for Volcano plot based on calibration threshold.}
\usage{
thrYaxis(thr, maxlogp)
}
\arguments{
\item{thr}{vector for calibration threshold}

\item{maxlogp}{numeric the maximum of -log(p-values)}
}
\value{
data.frame with the new labels and breaks for the Volcano Plot y-axis
}
\description{
Create a new y-axis for Volcano plot based on calibration threshold.
}
\examples{
thrYaxis(sort(runif(100)), 15)
}
