% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_output.R
\name{UrlStringdbGrah}
\alias{UrlStringdbGrah}
\title{Create an URL for string-db gene network from a vector of gene names.}
\usage{
UrlStringdbGrah(vector)
}
\arguments{
\item{vector}{a vector of gene names use for the gene networks
from string-db.}
}
\value{
a character, the url for string-db.org
}
\description{
Create an URL for string-db gene network from a vector of gene names.
}
