% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{matchMatrixBiofun}
\alias{matchMatrixBiofun}
\title{Check if genes in gene set match with a gene set list}
\usage{
matchMatrixBiofun(geneNames, biofun)
}
\arguments{
\item{geneNames}{vector of genes names (vector of charactors)}

\item{biofun}{matrix for gene set}
}
\description{
Check if genes in gene set match with a gene set list
}
